move examples to fed_kit_examples
