+++ android/app/src/main/java/org/eu/fedcampus/android_client/MainActivity.kt
@@ -9,10 +9,10 @@ import kotlinx.coroutines.MainScope
 import kotlinx.coroutines.launch
 import org.eu.fedcampus.android_client.databinding.ActivityMainBinding
 import org.eu.fedcampus.fed_kit.Train
-import org.eu.fedcampus.fed_kit.examples.cifar10.DATA_TYPE
-import org.eu.fedcampus.fed_kit.examples.cifar10.Float3DArray
-import org.eu.fedcampus.fed_kit.examples.cifar10.loadData
-import org.eu.fedcampus.fed_kit.examples.cifar10.sampleSpec
+import org.eu.fedcampus.fed_kit_examples.cifar10.DATA_TYPE
+import org.eu.fedcampus.fed_kit_examples.cifar10.Float3DArray
+import org.eu.fedcampus.fed_kit_examples.cifar10.loadData
+import org.eu.fedcampus.fed_kit_examples.cifar10.sampleSpec
 import org.eu.fedcampus.fed_kit_train.FlowerClient
 import org.eu.fedcampus.fed_kit_train.helpers.deviceId
 import org.eu.fedcampus.fed_kit_train.helpers.loadMappedFile
